Implemented addProductToCart, updateProductQuantity, removeProductFromCart, deleteCart methods.

+++ src/main/java/com/endava/mentorship2022/model/CartItem.java
@@ -16,7 +16,7 @@ public class CartItem {
     private long id;
 
     @Column(name = "quantity", nullable = false)
-    private int quantity;
+    private short quantity;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
